#2 Reformat to add Post files
