+++ ms_studenthelp/src/test/java/be/ecam/ms_studenthelp/ThreadUnitTest/CreateThreadTestForThreads.java
@@ -1,4 +1,4 @@
-package be.ecam.ms_studenthelp;
+package be.ecam.ms_studenthelp.ThreadUnitTest;
 
 import org.springframework.boot.json.JsonParser;
 import org.springframework.boot.json.JsonParserFactory;
@@ -12,10 +12,10 @@ import java.util.Map;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-public class CreateThreadTest {
-    public static String createThreadTest(MockMvc mockMvc, int port) throws Exception {
+public class CreateThreadTestForThreads {
+    public static String createThreadTestForThreads(MockMvc mockMvc, int port) throws Exception {
 
-        String json = "{\"title\":\"Test thread\",\"category\":\"Math\",\"tags\":[],\"firstPost\":{\"authorId\":\"d66b3f8c-2271-4afb-a348-e370effff\",\"content\":\"Fisrt postman\"}}";
+        String json = "{\"title\":\"Test thread\",\"category\":\"Math\",\"tags\":[],\"firstPost\":{\"authorId\":\"d66b3f8c-2271-4afb-a348-e370effff\",\"content\":\"First post\"}}";
 
         MvcResult result = mockMvc.perform(post("http://localhost:" + port + "/threads/")
                 .content(json)
@@ -27,6 +27,7 @@ public class CreateThreadTest {
         MockHttpServletResponse response = result.getResponse();
         JsonParser springParser = JsonParserFactory.getJsonParser();
         Map<String, Object> idResponse = springParser.parseMap(response.getContentAsString());
+        System.out.println(idResponse);
         return (String) idResponse.get("id");
     }
 }
