+++ ms_studenthelp/src/test/java/be/ecam/ms_studenthelp/ThreadUnitTest/UnitTestThreadsCreateContent.java
@@ -1,4 +1,4 @@
-package be.ecam.ms_studenthelp;
+package be.ecam.ms_studenthelp.ThreadUnitTest;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,21 +7,24 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.test.web.servlet.MockMvc;
 
+import static org.hamcrest.CoreMatchers.is;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
 @AutoConfigureMockMvc
-class UnitTestCheckDBStatus {
-
+class UnitTestThreadsCreateContent {
     private final int port = 8080;
     @Autowired
     private MockMvc mockMvc;
 
     @Test
-    public void unitTestCheckDBStatus() throws Exception {
-        this.mockMvc.perform(get("http://localhost:" + port + "/categories"))
-                .andExpect(status().isOk());
+    public void unitTestThreadsCreateContent() throws Exception {
+        String id = CreateThreadTestForThreads.createThreadTestForThreads(mockMvc, port);
+
+        this.mockMvc.perform(get("http://localhost:" + port + "/threads/" + id))
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.title", is("Test thread")));
     }
 }
-
